refactor(ai-gateway): derive model variants from catalog
