+++ apps/web/src/lib/code-reviews/core/model-variants.ts
@@ -1,8 +1,8 @@
-import { getModelVariants } from '@/lib/ai-gateway/providers/model-settings';
+import { getFallbackModelVariants } from '@/lib/ai-gateway/providers/variants';
 
 /** Returns the ordered list of thinking-effort variant names available for a model, or [] if the model has no variants. */
 export function getAvailableThinkingEfforts(modelSlug: string): string[] {
-  const variants = getModelVariants(modelSlug);
+  const variants = getFallbackModelVariants(modelSlug);
   return variants ? Object.keys(variants) : [];
 }
